requirements.txt
paho-mqtt==1.6.1
elastic-transport==8.4.0
elasticsearch==8.9.0
